+++ src/internal/FlowVideoView.tsx
@@ -10,13 +10,12 @@ import { useEditMode } from "./EditModeScope";
 import { useFlowCaretContext } from "./FlowCaretScope";
 import Color from "color";
 import { useBlockSize } from "./BlockSize";
-import { mdiAlert, mdiAlertOctagon, mdiLoading, mdiResizeBottomRight } from "@mdi/js";
+import { mdiResizeBottomRight } from "@mdi/js";
 import { useNativeEventHandler } from "./hooks/use-native-event-handler";
 import { useFlowEditorController } from "./FlowEditorControllerScope";
-import { useFlowLocale } from "../FlowLocaleScope";
-import Icon from "@mdi/react";
 import { useVerifiedImageUrl } from "./hooks/use-verified-image";
 import { useVideoSourceUrl } from "./hooks/use-video-source";
+import { UploadOverlay } from "./UploadOverlay";
 
 export const FlowVideoView = flowNode<FlowVideo>((props, outerRef) => {
     const { node, selection } = props;
@@ -35,7 +34,6 @@ export const FlowVideoView = flowNode<FlowVideo>((props, outerRef) => {
     const selected = selection === true;
     const editMode = useEditMode();
     const { native: nativeSelection } = useFlowCaretContext();
-    const locale = useFlowLocale();
 
     const rootClassName = clsx(
         classes.root,
@@ -68,7 +66,6 @@ export const FlowVideoView = flowNode<FlowVideo>((props, outerRef) => {
     const onError = useCallback(() => setBroken(!!videoUrl), [videoUrl]);
     const [wrapperElem, setWrapperElem] = useState<HTMLElement | null>(null);
     const [resizeStart, setResizeStart] = useState<{x:number,y:number,w:number,h:number}|null>(null);
-    const [uploadState, setUploadState] = useState<"not_available" | "in_progress" | "failed">();
 
     const onResizeStart = useCallback((e: React.MouseEvent<unknown>) => {
         if (wrapperElem && controller) {
@@ -155,7 +152,7 @@ export const FlowVideoView = flowNode<FlowVideo>((props, outerRef) => {
 
     const resizeOverlay = editMode && selected && controller && controller.isVideo() && (
         <>
-            {!uploadState && (
+            {!source.upload && (
                 <div className={classes.sizeProps}>
                     {Math.round(source.width * scale)} x {Math.round(source.height * scale)}<br/>
                     {(scale * 100).toFixed(1)}%
@@ -174,47 +171,6 @@ export const FlowVideoView = flowNode<FlowVideo>((props, outerRef) => {
             />
         </>
     );
-
-    const uploadOverlay = uploadState && (
-        <div className={classes.uploadOverlay}>
-            <div className={clsx(classes.uploadBox, classes[`upload_${uploadState}`])}>
-                <Icon
-                    className={classes.uploadIcon}
-                    path={UploadStateIcon[uploadState]}
-                    size={0.75}
-                    spin={uploadState === "in_progress" ? 1 : 0}
-                />
-                {locale[`video_upload_${uploadState}`]}
-            </div>
-        </div>
-    );
-
-    useEffect(() => {
-        if (!source.upload) {
-            setUploadState(undefined);
-        } else {
-            const promise = controller?.getUploadPromise(source.upload);
-            if (!promise) {
-                setUploadState("not_available");
-            } else {
-                let active = true;
-                setUploadState("in_progress");
-                promise.then(
-                    () => {
-                        if (active) {
-                            setUploadState(undefined);
-                        }
-                    },
-                    () => {
-                        if (active) {
-                            setUploadState("failed");
-                        }
-                    },
-                );
-                return () => void (active = false);
-            }
-        }
-    }, [source.upload, controller]);
 
     return (
         <span 
@@ -247,7 +203,7 @@ export const FlowVideoView = flowNode<FlowVideo>((props, outerRef) => {
                                     }
                                 />
                                 {stateOverlay}
-                                {uploadOverlay}
+                                <UploadOverlay uploadId={source.upload} controller={controller} type="video" />
                                 {resizeOverlay}
                             </>
                         )}
@@ -257,12 +213,6 @@ export const FlowVideoView = flowNode<FlowVideo>((props, outerRef) => {
         />
     );
 });
-
-const UploadStateIcon = {
-    in_progress: mdiLoading,
-    not_available: mdiAlert,
-    failed: mdiAlertOctagon,
-};
 
 const useStyles = createUseFlowStyles("FlowVideo", ({palette, typography}) => ({
     ...textStyles(palette, typography),
@@ -274,43 +224,6 @@ const useStyles = createUseFlowStyles("FlowVideo", ({palette, typography}) => ({
         outlineWidth: 2,
         outlineOffset: 4,
     },
-    uploadOverlay: {
-        position: "absolute",
-        top: 0,
-        bottom: 0,
-        left: 0,
-        right: 0,
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-        justifyContent: "center"
-    },
-    uploadBox: {
-        textIndent: "initial",
-        fontWeight: "normal",
-        backgroundColor: Color(palette.tooltip).fade(0.25).toString(),
-        color: palette.tooltipText,
-        fontFamily: typography.ui,
-        padding: 8,
-        border: `1px solid ${palette.tooltipText}`,
-        borderRadius: 4,
-        fontSize: "0.85rem",
-        "&$upload_not_available": {
-            backgroundColor: Color(palette.warning).fade(0.25).toString(),
-        },
-        "&$upload_failed": {
-            backgroundColor: Color(palette.error).fade(0.25).toString(),
-        },
-        display: "flex",
-        flexDirection: "row",
-        alignItems: "center"
-    },
-    uploadIcon: {
-        paddingRight: 8,
-    },
-    upload_in_progress: {},
-    upload_not_available: {},
-    upload_failed: {},
     selected: {
         outlineColor: palette.selection,
     },